rename demos
