+++ src/pages/countdown.js
@@ -4,10 +4,9 @@ import Link  from 'next/link';
 function Countdown() {
   return (
     <div>
-      <Link href="/countdown/game1">game1</Link>
-      <Link href="/countdown/game2">game2</Link>
-      <Link href="/countdown/game3">game3</Link>
-      <Link href="/countdown/game4">game4</Link>
+      <Link href="/countdown/p5GameDemo">p5.js Demo</Link>
+      <Link href="/countdown/3DPhysicsDemo">3D Physics Demo</Link>
+      <Link href="/countdown/3DAnimationDemo">3D Animation Demo</Link>
       <Leaderboard />
     </div>
   );
