+++ src/pages/countdown/3DAnimationDemo.js
@@ -1,0 +1,9 @@
+import MatchingGame from "components/3DAnimationDemo/matchingGame";
+
+function Animation3DDemo() {
+  return (
+    <MatchingGame />
+  )
+}
+
+export default Animation3DDemo;
