+++ src/pages/countdown/3DPhysicsDemo.js
@@ -1,0 +1,9 @@
+import PongGame from 'components/3DPhysicsDemo/pongGame'
+
+function Physics3DDemo() {
+  return (
+    <PongGame />
+  )
+}
+
+export default Physics3DDemo;
