--- src/pages/countdown/game1.js
@@ -1,18 +1,0 @@
-function Game1() {
-  return (
-  <>
-    <div>Glitch.me example</div>
-    <div style={{height: '600px', width: '100%'}}>
-      <iframe
-        src="https://glitch.com/embed/#!/embed/silver-cake-vinyl?path=README.md&previewSize=100"
-        title="silver-cake-vinyl on Glitch"
-        allow="geolocation; microphone; camera; midi; encrypted-media; xr-spatial-tracking; fullscreen"
-        allowFullScreen
-        style={{height: '100%', width: '100%', border: 0}}>
-      </iframe>
-    </div>
-  </>
-  );
-}
-
-export default Game1;
