--- src/pages/countdown/game2.js
@@ -1,23 +1,0 @@
-import React from "react";
-import { NextReactP5Wrapper } from "@p5-wrapper/next";
-
-const sketch = (p5) => {
-  p5.setup = () => p5.createCanvas(600, 400, p5.WEBGL);
-
-  p5.draw = () => {
-    p5.background(250);
-    p5.normalMaterial();
-    p5.push();
-    p5.rotateZ(p5.frameCount * 0.01);
-    p5.rotateX(p5.frameCount * 0.01);
-    p5.rotateY(p5.frameCount * 0.01);
-    p5.plane(100);
-    p5.pop();
-  };
-};
-
-function Game2() {
-  return <NextReactP5Wrapper sketch={sketch} />;
-}
-
-export default Game2;
