--- src/pages/countdown/game3.js
@@ -1,9 +1,0 @@
-import PongGame from 'components/game3/pongGame'
-
-function Game3() {
-  return (
-    <PongGame />
-  )
-}
-
-export default Game3;
