--- src/pages/countdown/game4.js
@@ -1,9 +1,0 @@
-import MatchingGame from "components/game4/matchingGame";
-
-function Game4() {
-  return (
-    <MatchingGame />
-  )
-}
-
-export default Game4;
