+++ src/pages/countdown/p5GameDemo.js
@@ -1,0 +1,8 @@
+import P5Sketch from "components/p5GameDemo/p5Sketch";
+
+
+function P5GameDemo() {
+  return <P5Sketch />;
+}
+
+export default P5GameDemo;
